Restart project (#10)
The original implementation of this project was a great starting point
to learn about Ansible. But I didn't keep the project up-to-date with
the changes that I made to my developer setup, and over the last two to
three years the difference became too large for this project to be
useful.

Instead of updating the repository, I am completely rewriting the
project based on lessons that I have learned and new inspirations that I
have found. It will still be based on Ansible, but this time use a
different approach and organization.