Create role for Homebrew
Homebrew is the missing package manager for macOS, and the preferred
tool to install software and even applications. A new role has been
added that installs Homebrew if it is missing, and updates it and all
packages to the latest version. This provides a good starting ground for
other roles that want to install something.

While Homebrew is also installed as part of the bootstrap script, adding
it explicitly as a role ensures that it is available to other roles no
matter how the project is used.