Require 1Password before running playbook (#28)
We plan to store sensitive values in 1Password, and use the 1Password
CLI to look them up at runtime. Because we are using two factor
authentication with 1Password, we cannot automate the CLI itself with
Ansible. The current workaround is thus to ensure that an active
session exists before running Ansible. This isn't great, but works
and makes the rest of the setup fairly easy.