Hotfix repository role (#41)
The repository role failed to execute with a rather weird error.
Apparently, the result object didn't have a `msg` key. This might be
caused by the loop, but I have no idea how to properly fix it. So we're
hotfixing the issue for the time being by ignoring all errors.